Finished EP26 and 27. Converted Hazel to static library. No more warnings, and created Render-API specific context class for OpenGL etc. Note: For the documentation/comments on previous methods that are moved to rendering, are now archived inside OpenGLContext.h.

+++ Hazel/src/Platform/Windows/WindowsInput.cpp
@@ -27,13 +27,15 @@ namespace Hazel {
 		return state == GLFW_PRESS;
 	}
 
-	std::pair<float, float> WindowsInput::GetMousePositionImpl() const
-	{
+	std::pair<float, float> WindowsInput::GetMousePositionImpl() const {
 
 		auto window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
-		return std::pair<float, float>(xpos, ypos);
+
+		// need to cast double, to substitute to float, to prevent warning 
+		// not sure why float returned instead of double yet
+		return std::pair<float, float>((float)xpos, (float)ypos);
 	}
 
 	float WindowsInput::GetMouseXImpl() const {
